change chancecard to use usableitem
