--- src/main/kotlin/jp/trap/conqest/game/ChanceCard.kt
@@ -1,19 +1,0 @@
-package jp.trap.conqest.game
-
-import org.bukkit.ChatColor
-import org.bukkit.Material
-import org.bukkit.inventory.ItemStack
-
-class ChanceCard {
-    companion object {
-        fun createChanceCard(): ItemStack {
-            val chanceCard = ItemStack(Material.POISONOUS_POTATO)
-            val meta = chanceCard.itemMeta
-            if (meta != null) {
-                meta.setDisplayName("${ChatColor.GOLD}チャンスカード")
-                chanceCard.setItemMeta(meta)
-            }
-            return chanceCard
-        }
-    }
-}
